requirements_optimized.txt
# Core dependencies
requests>=2.31.0
aiohttp>=3.8.0
asyncio>=3.4.3
structlog>=23.1.0
python-dotenv>=1.0.0
openai>=1.0.0
aiofiles>=23.1.0

# Web server
flask>=2.0.0
flask-cors>=3.0.0

# Audio processing
soundfile>=0.12.1
numpy>=1.21.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6

# Original dependencies
# (keep existing ones from your setup) 